requirements.txt
python-dotenv>=1.0.0
langchain-core>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=0.0.5
slack-bolt>=1.18.0
slack-sdk>=3.21.0 